SwingController introduced as required in MVC. MandelbrotApp just combining Model, View and Controller.
